% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{add_eng_uk_tots}
\alias{add_eng_uk_tots}
\title{Add England and UK Totals}
\usage{
add_eng_uk_tots(
  df,
  of_colmn,
  by_var = NULL,
  cars.areas = NULL,
  cars.names = NULL
)
}
\arguments{
\item{df}{Data frame to which totals will be added.}

\item{of_colmn}{Column name to sum.}

\item{by_var}{Optional grouping variable.}

\item{cars.areas}{Optional vector of car area values.}

\item{cars.names}{Optional vector of car names.}
}
\value{
Data frame with England and UK totals added.
}
\description{
This function adds England and UK total rows to a data frame, summing up values as needed.
}
