% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cars.factfile.direct}
\alias{cars.factfile.direct}
\title{CARs Factfile Directory}
\format{
A character vector representing the file path.
}
\usage{
cars.factfile.direct
}
\description{
This is the path to the CARs factfile directory.
}
\examples{
cars.factfile.direct
}
\keyword{datasets}
