% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ceh.col.pallette}
\alias{ceh.col.pallette}
\title{CEH Color Palette}
\format{
A character vector of hex color codes.
}
\usage{
ceh.col.pallette
}
\description{
A color palette used for CEH land cover types.
}
\examples{
ceh.col.pallette
}
\keyword{datasets}
