% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ceh.concatted1.habtype}
\alias{ceh.concatted1.habtype}
\title{Concatenated Habitat Types}
\format{
A character vector of habitat types.
}
\usage{
ceh.concatted1.habtype
}
\description{
A vector containing concatenated habitat types from the CEH dataset.
}
\examples{
ceh.concatted1.habtype
}
\keyword{datasets}
