% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ceh.full.habtype}
\alias{ceh.full.habtype}
\title{Full Habitat Types from CEH}
\format{
A factor with levels corresponding to habitat types.
}
\usage{
ceh.full.habtype
}
\description{
A vector containing full habitat types from the CEH dataset.
}
\examples{
ceh.full.habtype
}
\keyword{datasets}
