% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{colour.brks}
\alias{colour.brks}
\title{Generate Color Breaks}
\usage{
colour.brks(lims, n = 5)
}
\arguments{
\item{lims}{Numeric vector of length 2. Specifies the minimum and maximum limits for the breaks.}

\item{n}{Integer. The number of breaks to generate. Default is 5.}
}
\value{
A numeric vector of breaks.
}
\description{
This function generates a sequence of evenly spaced numeric breaks within the specified limits.
}
\examples{
lims <- c(1, 10)
colour.brks(lims, n = 5)

}
