% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{combine_sf_batches}
\alias{combine_sf_batches}
\title{Combine sf objects in batches}
\usage{
combine_sf_batches(sf_list, batch_size = 1000)
}
\arguments{
\item{sf_list}{A list of sf objects to be combined.}

\item{batch_size}{An integer specifying the size of each batch. Defaults to 1000.}
}
\value{
A list of combined sf objects, one for each batch.
}
\description{
This function takes a list of sf objects and combines them into a single sf object
in batches to optimize performance. It is especially useful when working with large
lists of spatial features.
}
\examples{
\dontrun{
combined_sf <- combine_sf_batches(sf_list, batch_size = 500)
}
}
