% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{countries.names}
\alias{countries.names}
\title{Country Names}
\format{
A character vector with the names of countries.
}
\usage{
countries.names
}
\description{
A vector containing the names of countries (England, Northern Ireland, Scotland, Wales).
}
\examples{
countries.names
}
\keyword{datasets}
