% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{create_ordered_table}
\alias{create_ordered_table}
\title{Create Ordered Table}
\usage{
create_ordered_table(
  data,
  row_var,
  col_var,
  type = c("frequency", "proportion"),
  round_digits = 3
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{row_var}{The variable name (as a string) to be used for the rows.}

\item{col_var}{The variable name (as a string) to be used for the columns.}

\item{type}{The type of table to create. It can be either "frequency" or "proportion". Default is "frequency".}

\item{round_digits}{The number of digits to round the proportions to. Default is 3.}
}
\value{
A table with ordered rows and margins added.
}
\description{
This function creates a table based on the provided row and column variables, and orders
the rows by their total sum in descending order. You can choose to create a frequency table
or a proportion table. The table can also be rounded to a specified number of digits.
}
