% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{england.car.names}
\alias{england.car.names}
\title{England CARs Names}
\format{
A character vector with the names of England CARs regions.
}
\usage{
england.car.names
}
\description{
A vector containing the names of England CARs regions.
}
\examples{
england.car.names
}
\keyword{datasets}
