% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{find.lims}
\alias{find.lims}
\title{Find Limits for Variable Scaling}
\usage{
find.lims(
  var,
  quant.weights = rep(1, length(var)),
  consider = rep(TRUE, length(var)),
  quant.prob = 0.98,
  sd.mult = 3
)
}
\arguments{
\item{var}{Numeric vector. The variable for which limits are being calculated.}

\item{quant.weights}{Numeric vector. Weights for each observation in \code{var}, used for calculating the weighted quantile. Default is equal weights.}

\item{consider}{Logical vector. Indicator of which elements in \code{var} should be considered. Default is all elements.}

\item{quant.prob}{Numeric. Probability for the quantile calculation. Default is 0.98.}

\item{sd.mult}{Numeric. Multiple of the standard deviation to add to the mean. Default is 3.}
}
\value{
Numeric. The upper limit for the variable scaling.
}
\description{
This function computes an upper limit for scaling variables for display purposes, useful for squishing variable scales to a specified quantile or standard deviation.
The limit is determined as the minimum of two values: a mean plus a multiple of standard deviation or a weighted quantile.
}
\examples{
\dontshow{if (requireNamespace("Hmisc", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
var <- rnorm(100)
find.lims(var)
\dontshow{\}) # examplesIf}
}
