% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{gis.wd}
\alias{gis.wd}
\title{GIS Working Directory}
\format{
A character vector representing the file path.
}
\usage{
gis.wd
}
\description{
This is the path to the GIS working directory used across projects.
}
\examples{
gis.wd
}
\keyword{datasets}
