% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{load_lcm}
\alias{load_lcm}
\title{Load LCM data as terra SpatRasters}
\usage{
load_lcm(lcm.directs, years, resolution = "25")
}
\arguments{
\item{lcm.directs}{A data frame containing file paths and year metadata}

\item{years}{Vector of years to include}

\item{resolution}{Character. One of "10", "20", "25"}
}
\value{
A list with:
\describe{
\item{gb}{SpatRaster (years as layers)}
\item{ni}{SpatRaster (years as layers)}
}
}
\description{
Loads GB and NI LCM rasters for selected years and resolution,
returning SpatRaster time stacks.
}
