% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{load_lcm_year}
\alias{load_lcm_year}
\title{Load LCM rasters using terra}
\usage{
load_lcm_year(year, lcm.directs)
}
\arguments{
\item{year}{Numeric or character. Year to extract.}

\item{lcm.directs}{Data frame containing:
\itemize{
\item year
\item gb.25 (GB raster sources)
\item ni.25 (NI raster sources)
}}
}
\value{
A list with:
\describe{
\item{gb}{SpatRaster for Great Britain}
\item{ni}{SpatRaster for Northern Ireland}
}
}
\description{
Loads GB and NI LCM rasters for a specified year using terra,
standardises sea values, and returns a structured list.
}
\examples{
\dontrun{
lcm <- load_lcm_year_terra(2015, lcm.directs)
}
}
