% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{main.wd}
\alias{main.wd}
\title{Main Working Directory}
\format{
A character vector representing the file path.
}
\usage{
main.wd
}
\description{
This is the path to the main working directory used across projects.
}
\examples{
main.wd
}
\keyword{datasets}
