% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{pad.lim}
\alias{pad.lim}
\title{Pad Range of Numeric Vector with Percentage Padding}
\usage{
pad.lim(x, map.pad = 0.05)
}
\arguments{
\item{x}{Numeric vector. The data for which the range is to be padded.}

\item{map.pad}{Numeric. The percentage of the range to add as padding to both ends. Default is 0.05 (5\%).}
}
\value{
Numeric vector of length 2. The padded range, with the specified percentage added to both ends.
}
\description{
This function pads the range of a numeric vector by a specified percentage. The padding is added to both the minimum and maximum values of the vector.
}
\examples{
pad.lim(c(10, 15, 20, 25))
pad.lim(c(50, 60, 70, 80), map.pad = 0.1)

}
