% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{print_with_commas}
\alias{print_with_commas}
\title{Format Vector with Commas}
\usage{
print_with_commas(vec)
}
\arguments{
\item{vec}{Character vector.}
}
\value{
Formatted string.
}
\description{
This function formats a vector of strings with commas and "and" for readability.
}
