% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{replace_rounded_zeros_with_string}
\alias{replace_rounded_zeros_with_string}
\title{Replace Rounded Zeros with a String}
\usage{
replace_rounded_zeros_with_string(data, precision = 3)
}
\arguments{
\item{data}{A numeric vector or matrix.}

\item{precision}{The number of decimal places to round the data. Default is 3.}
}
\value{
A vector or matrix with values replaced by a string if they are greater than 0
but round to 0, and the rest rounded to the specified precision.
}
\description{
This function rounds numeric data to a specified precision and replaces values
greater than 0 but rounded to 0 with a string indicating a value smaller than the threshold.
}
\examples{
data <- c(0.0004, 0.00003, 0.5, 1)
replace_rounded_zeros_with_string(data, precision = 3)

}
