% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{theme_map}
\alias{theme_map}
\title{Custom ggplot2 Map Theme}
\usage{
theme_map(
  leg.tit.size = 8,
  legend.position = "bottom",
  include.title = TRUE,
  ...
)
}
\arguments{
\item{leg.tit.size}{Numeric. Size of the legend title text. Default is 8.}

\item{legend.position}{Character or numeric vector. Position of the legend.
Can be one of "none", "left", "right", "bottom", "top", or a numeric vector
of length two. Default is "bottom".}

\item{include.title}{Logical. If \code{TRUE}, includes plot title and subtitle with
the default text size. If \code{FALSE}, both title and subtitle are removed. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 \code{theme} object.
}
\description{
This function provides a customized \code{ggplot2} theme designed for map visualizations.
It includes options to control the inclusion of titles and subtitles, and allows
customization of legend size and position.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(x = displ, y = hwy)) +
    ggplot2::geom_point() +
    theme_map(leg.tit.size = 10, legend.position = "right", include.title = FALSE)
}

}
