% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{this.andAll.tss}
\alias{this.andAll.tss}
\title{Treescape Names with All Option}
\format{
A character vector with the names.
}
\usage{
this.andAll.tss
}
\description{
A vector containing the LCM names for treescapes including an 'All Treescapes' option.
}
\examples{
this.andAll.tss
}
\keyword{datasets}
