% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ts.abbrv}
\alias{ts.abbrv}
\title{Treescape Abbreviations}
\format{
A character vector with the abbreviations of treescapes.
}
\usage{
ts.abbrv
}
\description{
A vector containing the abbreviations of treescapes.
}
\examples{
ts.abbrv
}
\keyword{datasets}
