% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ts.andAll.lcm.names}
\alias{ts.andAll.lcm.names}
\title{LCM Names for Treescapes with All Option}
\format{
A character vector with the LCM names.
}
\usage{
ts.andAll.lcm.names
}
\description{
A vector containing the LCM names for treescapes including an 'All Treescapes' option.
}
\examples{
ts.andAll.lcm.names
}
\keyword{datasets}
