% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ts.andAll.nice.names}
\alias{ts.andAll.nice.names}
\title{Nice Names for Treescapes with All Option}
\format{
A character vector with the nice names.
}
\usage{
ts.andAll.nice.names
}
\description{
A vector containing the nice names of treescapes including an 'All Treescapes' option.
}
\examples{
ts.andAll.nice.names
}
\keyword{datasets}
