% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ts.lcm.names}
\alias{ts.lcm.names}
\title{LCM Names for Treescapes}
\format{
A character vector with the LCM names.
}
\usage{
ts.lcm.names
}
\description{
A vector containing the LCM names for treescapes.
}
\examples{
ts.lcm.names
}
\keyword{datasets}
