% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ts.names}
\alias{ts.names}
\title{Treescape Names}
\format{
A character vector with the names of treescapes.
}
\usage{
ts.names
}
\description{
A vector containing the names of treescapes.
}
\examples{
ts.names
}
\keyword{datasets}
