% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ts.nice.names}
\alias{ts.nice.names}
\title{Nice Names for Treescapes}
\format{
A character vector with the nice names.
}
\usage{
ts.nice.names
}
\description{
A vector containing the nice names of treescapes.
}
\examples{
ts.nice.names
}
\keyword{datasets}
