% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{Umake.target.hexgrid}
\alias{Umake.target.hexgrid}
\title{Generate an Efficient Target Hexagonal Grid}
\usage{
Umake.target.hexgrid(
  master.grid.mask = countries,
  grid.size = c(hexdist.h, hexdist.v),
  target
)
}
\arguments{
\item{master.grid.mask}{sf object. A spatial feature representing the area over which the initial grid is generated. Default is \code{countries}.}

\item{grid.size}{Numeric vector of length 2. Specifies the horizontal and vertical distances between the centers of adjacent hexagons.
Example: \code{c(hexdist.h, hexdist.v)}.}

\item{target}{sf object. The target area to intersect with the hexagonal grid.}
}
\value{
An \code{sf} object representing the hexagonal grid intersected with the target area. The object contains the grid ID and the area of each hexagon in hectares.
}
\description{
This function creates a hexagonal grid over a specified target area by intersecting it with a master grid mask.
The resulting hexagonal grid is associated with grid IDs and calculates the area of each hexagon in hectares.
This optimized version reduces redundant operations and improves spatial computation efficiency.
}
