% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{colour.brks}
\alias{colour.brks}
\title{Generate Color Breaks with Significant Figures}
\usage{
colour.brks(lims, n = 5, sig_figs = 3)
}
\arguments{
\item{lims}{Numeric vector of length 2. Specifies the minimum and maximum limits for the breaks.}

\item{n}{Integer. The number of breaks to generate. Default is 5.}

\item{sig_figs}{Integer. The number of significant figures to which the breaks should be rounded. Default is 3.}
}
\value{
A numeric vector of breaks rounded to the specified number of significant figures.
}
\description{
This function generates a sequence of evenly spaced numeric breaks within the specified limits,
rounded to a specified number of significant figures.
}
\examples{
lims <- c(1, 10)
colour.brks(lims, n = 5, sig_figs = 3)

}
