% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{colour.brks}
\alias{colour.brks}
\title{Generate Color Breaks with Rounding}
\usage{
colour.brks(lims, n = 5, round_to = 100, just_pretty = T)
}
\arguments{
\item{lims}{Numeric vector of length 2. Specifies the minimum and maximum limits for the breaks.}

\item{n}{Integer. The number of breaks to generate. Default is 5.}

\item{round_to}{Numeric. The value to which the breaks should be rounded (e.g., 10 for tens, 100 for hundreds). Default is 1 (no rounding).}
}
\value{
A numeric vector of rounded breaks.
}
\description{
This function generates a sequence of numeric breaks within the specified limits
and rounds them to the nearest specified value.
}
\examples{
lims <- c(1, 1000)
colour.brks(lims, n = 5, round_to = 100)

}
