% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{st_erase}
\alias{st_erase}
\title{Erase Geometries from a Spatial Object}
\usage{
st_erase(x, y)
}
\arguments{
\item{x}{An \code{sf} object. The spatial object from which geometries will be removed.}

\item{y}{An \code{sf} object. The spatial object containing geometries to be erased from \code{x}.}
}
\value{
An \code{sf} object. The resulting spatial object after erasing the geometries in \code{y} from \code{x}.
}
\description{
A helper function that erases all of y from x. This function removes all geometries in one spatial object (\code{y}) from another spatial object (\code{x}).
It performs a spatial difference operation, effectively erasing the area covered by \code{y} from \code{x}.
}
