% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_utilities_functions.R
\name{st_first.spatial.curation}
\alias{st_first.spatial.curation}
\title{Perform Initial Spatial Curation on Polygon Data}
\usage{
st_first.spatial.curation(
  x,
  tolerance = 10,
  tiny.buff = 1e-04,
  smallest.hole = 5000
)
}
\arguments{
\item{x}{An \code{sf} object of type \code{POLYGON} or \code{MULTIPOLYGON}. The spatial data to be curated.}

\item{tolerance}{Numeric. The tolerance level for geometry simplification. Higher values result in more simplification. Default is 10.}

\item{tiny.buff}{Numeric. The distance (in the CRS units) to buffer polygons to address tiny gaps. Default is 0.0001.}

\item{smallest.hole}{Numeric. The minimum area of holes to be removed. Currently not used in this implementation but can be added for additional filtering.}
}
\value{
An \code{sf} object of type \code{POLYGON} or \code{MULTIPOLYGON}. The curated spatial data after applying the transformations.
}
\description{
This function applies a series of spatial operations to clean and simplify polygon data. It transforms the coordinate reference system, simplifies the geometry, buffers to address tiny gaps, and makes the geometries valid.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
